+++ gcloud-java-compute/src/test/java/com/google/gcloud/compute/ZoneTest.java
@@ -23,13 +23,14 @@ import com.google.gcloud.compute.Zone.MaintenanceWindow;
 
 import org.junit.Test;
 
+import java.math.BigInteger;
 import java.util.List;
 
 public class ZoneTest {
 
   private static final ZoneId ZONE_ID = ZoneId.of("project", "zone");
   private static final RegionId REGION_ID = RegionId.of("project", "region");
-  private static final Long ID = 42L;
+  private static final BigInteger ID = BigInteger.valueOf(42L);
   private static final String CREATION_TIMESTAMP = "2016-01-20T04:39:00.210-08:00";
   private static final String DESCRIPTION = "description";
   private static final Zone.Status STATUS = Zone.Status.DOWN;
